requirements.txt
Flask==2.1.2
line-bot-sdk==2.2.1
gunicorn==20.1.0
dj-database-url==0.5.0
psycopg2==2.9.3